Move leadership code into internal/k8s
This moves the code into k8s.
