+++ internal/k8s/k8s.go
@@ -215,9 +215,6 @@ func (c *Client) Errorf(svc *v1.Service, kind, msg string, args ...interface{}) 
 	c.events.Eventf(svc, v1.EventTypeWarning, kind, msg, args...)
 }
 
-// ClientCoreV1 returns the embeded k8s client from Client.
-func (c *Client) ClientCoreV1() v1core.CoreV1Interface { return c.client.CoreV1() }
-
 // NodeHasHealthyEndpoint return true if this node has at least one healthy endpoint.
 func NodeHasHealthyEndpoint(eps *v1.Endpoints, node string) bool {
 	ready := map[string]bool{}
